@@ -18,17 +18,20 @@
     "build": "rollup -c",
     "prepack": "npm run test && npm run build"
   },
+  "dependencies": {
+    "@faker-js/faker": "^9.5.1"
+  },
   "devDependencies": {
     "@rollup/plugin-commonjs": "^23.0.2",
     "@rollup/plugin-node-resolve": "^15.0.1",
-    "@rollup/plugin-typescript": "^9.0.2",
-    "rollup":"3.29.5",
+    "@rollup/plugin-typescript": "^12.1.2",
+    "rollup": "3.29.5",
     "rollup-plugin-dts": "^5.0.0",
     "rollup-plugin-peer-deps-external": "^2.2.4",
     "rollup-plugin-terser": "^7.0.2",
-    "typescript": "^4.8.4"
+    "typescript": "^5.8.3"
   },
   "overrides": {
-    "rollup":"3.29.5"
+    "rollup": "3.29.5"
   }
 }
