@@ -17,7 +17,7 @@ const { resolveAwsV4Credentials, addAwsV4Interceptor } = require('./awsv4auth-he
 const { addDigestInterceptor } = require('./digestauth-helper');
 const prepareGqlIntrospectionRequest = require('./prepare-gql-introspection-request');
 const { prepareRequest } = require('./prepare-request');
-const interpolateVars = require('./interpolate-vars');
+const { interpolateVars } = require('@usebruno/common');
 const { makeAxiosInstance } = require('./axios-instance');
 const { cancelTokens, saveCancelToken, deleteCancelToken } = require('../../utils/cancel-token');
 const { uuid, safeStringifyJSON, safeParseJSON, parseDataFromResponse, parseDataFromRequest } = require('../../utils/common');
